Move patch functionality for apply into its own file.
